Add benchmark of non-RAC type with RAC implementation
This relaxes the RAC requirement on Zip2RAC to allow bidirectional
collections as well as a benchmark that zips a long string with a variable
length range. The results show the high cost of this implementation, as the
string's count needs to be calculated directly even when it is much, much
larger than the paired range.