Customize implementation of MultiSets.CountedSet.count
The custom implementation of count adds the multiplicities of the elements together, which is far more efficient than the default approach of iterating through the set manually.

underestimatedCount has also been customized: it exposes the count of the underlying dictionary, which is guaranteed to be less than or equal to the cardinality of the counted set itself. This may be used to avoid trapping for sets with very large multiplicities.