Disable modules that haven’t gone through API review yet
These are disabled on the package manifest level.

To locally reenable them, Package.swift needs to be manually edited to enable COLLECTIONS_ENABLE_UNSTABLE_MODULES.

To mark a module as ready for production use, it has to be defined using `registerLibrary` instead of `registerUnstableLibrary`. This can only be done once the module has gone through API review. (Or if it becomes an underscored — i.e., explicitly unstable — module.)