Let deque have its capacity shrunk
Deque is often used as a buffer and buffers can be subject to "spiky"
sizes; many elements may be buffered before being removed which can
result in applications holding on to more memory than desired.

Since the capacity of deque is an implementation detail, one way to reduce
the size of the deque is asking it to shrink itself. This change adds
a `srinkCapacity` method which will reduce the storage capacity of a
deque if:
- there are fewer elements in the deque than the target capacity
- the target capacity is less than the previous requested capacity (the
  actual allocated capacity may be larger than the requested)