Remove workaround for SR-14663
Our updated minimum toolchain requirement is Swift 5.7, and the stdlib in that version has the fix included.

We’ve seen reports that the bug may not deploy back in all cases to earlier OS releases, but it still seems worth trying to remove the workaround, as it has some undesirable runtime implications.