Revive most Container protocols
We cannot express the right subscript requirements yet; make do by requiring hidden members that return (unsafe, escaping) pointers, and provide a default subscript that’s based on unsafe addressors.

This is very unsatisfying, but currently there is no way for a protocol to require subscripts with the right semantics.