Add `async` overloads of all `OutputSpan` providing APIs
## Motivation

It is currently not possible to use the new `OutputSpan` based APIs in asynchronous contexts. In particular, it is not possible to initialize or append to a `Unique/RigidArray` by reading asynchronously from the network or a file.

## Modifications

This PR adds new overloads to all the existing APIs that provide an `OutputSpan` to take an asynchronous closure. The PR also adds tests for all those new methods to show that they are properly working.

## Result

We can now interact with these output spans asynchronously.